Update mp prototype to support multiple lobbies, separate Lobby code from Multiplayer.js
This commit introduces multiple lobby support to the multiplayer prototype and separates the Lobby functionality from the Multiplayer.js file.
