+++ KeyCtrl/package.json
@@ -26,10 +26,11 @@
     "react-router-dom": "^6.2.1",
     "react-scripts": "^4.0.3",
     "react-spring": "^9.3.0",
+    "request": "^2.88.2",
     "request-promise": "^4.2.6",
+    "socket.io-client": "^2.3.0",
     "styled-components": "^5.3.3",
-    "web-vitals": "^1.1.2",
-     "socket.io-client": "^2.3.0"
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
