+++ KeyCtrl/src/components/MultiplayerPage/Multiplayer.js
@@ -1,8 +1,7 @@
-import TextField from "@material-ui/core/TextField"
-import React, { useEffect, useRef, useState } from "react"
-import io from "socket.io-client"
+import React, { useState } from "react"
 import styled from 'styled-components'
 import '../../styles/MultiplayerPage.css'
+import { Lobby } from './Lobby.js'
 
 const Button = styled.button`
   background: #3D3883;
@@ -19,81 +18,22 @@ const Button = styled.button`
   }
 `
 
-const Multiplayer = () => {
-
+const Multiplayer = () => {  
+  //Set lobby join state and update during button press
+  const [joinLobby, inLobby] = useState(false)
+  var lobbyID = 1
   function findMatch() {
-    console.log("Find match clicked");
+    //ADD CODE TO GENERATE LOBBY ID
+    inLobby(true)
   }
-
-  const [state, setState] = useState({ message: "", name: "" })
-  const [chat, setChat] = useState([])
-
-  const socketRef = useRef()
-
-  useEffect(
-    () => {
-      socketRef.current = io.connect("151.197.236.81:4000")
-      socketRef.current.on("message", ({ name, message }) => {
-        setChat([...chat, { name, message }])
-      })
-      return () => socketRef.current.disconnect()
-    },
-    [chat]
-  )
-
-  const onTextChange = (e) => {
-    setState({ ...state, [e.target.name]: e.target.value })
-  }
-
-  const onMessageSubmit = (e) => {
-    const { name, message } = state
-    socketRef.current.emit("message", { name, message })
-    e.preventDefault()
-    setState({ message: "", name })
-  }
-
-  const renderChat = () => {
-    return chat.map(({ name, message }, index) => (
-      <div key={index}>
-        <h3>
-          {name}: <span>{message}</span>
-        </h3>
-      </div>
-    ))
-  }
-
 
   return (
     <div>
       <div className='multiplayer-base'>
         <div>Multiplayer</div>
-        <Button onClick={findMatch} >Find Match</Button>
+        {joinLobby ? null : <Button onClick={findMatch} >Find Match</Button>}
+        {joinLobby ? <Lobby lobbyID={lobbyID}/> : null}
       </div>
-
-      <div className="card">
-        <form onSubmit={onMessageSubmit}>
-          <h1>Messenger</h1>
-          <div className="name-field">
-            <TextField name="name" onChange={(e) => onTextChange(e)} value={state.name} label="Name" />
-          </div>
-          <div>
-            <TextField
-              name="message"
-              onChange={(e) => onTextChange(e)}
-              value={state.message}
-              id="outlined-multiline-static"
-              variant="outlined"
-              label="Message"
-            />
-          </div>
-          <button>Send Message</button>
-        </form>
-        <div className="render-chat">
-          <h1>Chat Log</h1>
-          {renderChat()}
-        </div>
-      </div>
-      
     </div>
   )
 }
